EnvWeeeView timestamp

+++ src/main/java/pl/com/pt4q/product_manager/modules/environment/data/weee/EnvWeeeEntity.java
@@ -32,9 +32,9 @@ public class EnvWeeeEntity {
     @ManyToOne
     private UnitEntity netWeightUnit;
 
-    private Integer itemHeight;
-    private Integer itemLength;
-    private Integer itemDepth;
+    private Double itemHeight;
+    private Double itemLength;
+    private Double itemDepth;
     @ManyToOne
     private UnitEntity itemLengthUnit;
